Upgrade to Devise 4.9.0.alpha with Turbo support
Remove the previous custom code to make it work with Turbo, as that can
interfere with Devise itself and cause other issues.

Note that I enforced responders v3.1 in the Gemfile as it's necessary to
work with Devise + Turbo, but we could've just updated it via `bundle
update responders` too. (this can't be enforced by Devise itself due to
compatibility with older Rails versions yet.)